Add files via upload
First Draft of CAD Files (Arduino enclosure and Finger mount) 
Finger mount can be secured on arduino enclosure for transportation and storage and then in use may be detached and placed on a surface for use.  Finger will have a squishy pad that is used, as finger tremors the opening of the finger mount will be shut off and opened to light.